+++ graphql-dgs-spring-boot-oss-autoconfigure/build.gradle.kts
@@ -23,6 +23,7 @@ dependencies {
 
     compileOnly("com.github.ben-manes.caffeine:caffeine")
     compileOnly("io.micrometer:micrometer-core")
+    compileOnly("io.projectreactor:reactor-core")
 
     testImplementation("org.springframework.boot:spring-boot-starter-web")
     testImplementation("com.github.ben-manes.caffeine:caffeine")
