+++ graphql-dgs/dependencies.lock
@@ -61,6 +61,9 @@
         "com.netflix.graphql.dgs:graphql-error-types": {
             "project": true
         },
+        "io.projectreactor:reactor-core": {
+            "locked": "3.4.10"
+        },
         "org.jetbrains.kotlin:kotlin-bom": {
             "locked": "1.5.32"
         },
@@ -462,6 +465,9 @@
         "org.jetbrains.kotlinx:kotlinx-coroutines-jdk8": {
             "locked": "1.3.8"
         },
+        "org.jetbrains.kotlinx:kotlinx-coroutines-test": {
+            "locked": "1.3.8"
+        },
         "org.springframework.boot:spring-boot-dependencies": {
             "locked": "2.3.12.RELEASE"
         },
@@ -558,6 +564,9 @@
             "locked": "1.3.8"
         },
         "org.jetbrains.kotlinx:kotlinx-coroutines-jdk8": {
+            "locked": "1.3.8"
+        },
+        "org.jetbrains.kotlinx:kotlinx-coroutines-test": {
             "locked": "1.3.8"
         },
         "org.slf4j:slf4j-api": {
