+++ graphql-dgs/src/main/kotlin/com/netflix/graphql/dgs/internal/ReactiveDataFetcherResultProcessor.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2021 Netflix, Inc.
+ * Copyright 2022 Netflix, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,10 +14,9 @@
  * limitations under the License.
  */
 
-package com.netflix.graphql.dgs.reactive.internal
+package com.netflix.graphql.dgs.internal
 
 import com.netflix.graphql.dgs.DgsDataFetchingEnvironment
-import com.netflix.graphql.dgs.internal.DataFetcherResultProcessor
 import reactor.core.publisher.Flux
 import reactor.core.publisher.Mono
 import java.lang.IllegalArgumentException
